deprecate utils/signatures in favor of utils/signature
